refactor: Convert data loading from JS modules to fetch API
