+++ README.md
@@ -1,10 +1,13 @@
 # checklist
 [ ] 캐러셀 보완
-    - 마지막 페이지는 남은 카드 갯수만큼만 움직이도록 함
-[ ] 메인 페이지 개선
-    [ ] 데이터를 json 형태로 저장
+    [ ] 유한 캐러셀
+        - 맨 끝에서 앞으로 갈 때 2개만 움직이지 않고 5개씩 가다가 맨 앞에서 2개만 움직이도록 함
+    [ ] 무한 캐러셀
+        - 마지막 페이지는 남은 카드 갯수만큼만 움직이도록 함
+[x] 메인 페이지 개선
+    [x] 데이터를 json 형태로 저장
         - json 파일을 만들어서 요청하는 과정 찾아보기
-    [ ] fetch 요청으로 데이터를 화면에 노출
+    [x] fetch 요청으로 데이터를 화면에 노출
 [ ] 기능 추가
     [ ] 좋아요 기능
         - 눌리기 전에는 비어있고 눌리면 색이 채워지도록 함
@@ -13,7 +16,7 @@
         - 시간 나면 해보기
 [ ] vite 도구로 개발환경 구성 및 적용
     - vite 개발 환경으로 만든 새 폴더에 src를 옮겨서 build 후 작업 폴더로 src를 이전
-[ ] 코드 리팩토링
+[x] 코드 리팩토링
     - 클래스 명 정리
     - HTML 길이 줄이고 JS로 역할 위임
 [ ] TypeScript 적용
