+++ data/footerData.json
@@ -1,0 +1,18 @@
+{
+  "socialLogos": [
+    { "src": "./assets/img_footer/facebook.png", "alt": "facebook_logo" },
+    { "src": "./assets/img_footer/instagram.png", "alt": "instagram_logo" },
+    { "src": "./assets/img_footer/x.png", "alt": "x_logo" },
+    { "src": "./assets/img_footer/youtube.png", "alt": "youtube_logo" }
+  ],
+  "links": ["화면 해설", "고객 센터", "기프트카드", "미디어 센터", "투자 정보(IR)", "입사 정보", "이용 약관", "개인정보", "법적 고지", "쿠키 설정", "회사 정보", "문의하기"],
+  "info": [
+    "넷플릭스서비시스코리아 유한회사 통신판매업신고번호: 제2018-서울종로-0426호 전화번호: 00-308-321-0161 (수신자 부담)",
+    "대표: 레지널드 숀 톰프슨",
+    "이메일 주소: korea@netflix.com",
+    "주소: 대한민국 서울특별시 종로구 우정국로 26, 센트로폴리스 A동 20층 우편번호 03161",
+    "사업자등록번호: 165-87-00119",
+    "클라우드 호스팅: Amazon Web Services Inc.",
+    "공정거래위원회 웹사이트"
+  ]
+}
