+++ data/heroData.json
@@ -1,0 +1,17 @@
+{
+  "background": { "src": "./assets/img_hero/overlay.jpg", "alt": "폭군의 셰프 사진" },
+  "titleImg": { "src": "./assets/img_hero/overname.png", "alt": "폭군의 셰프" },
+  "ranking": {
+    "badge": { "src": "./assets/img_hero/top10.jpg", "alt": "top10_logo" },
+    "text": "오늘 시리즈 순위 2위"
+  },
+  "description": "시간을 거슬러 과거로 가게 된 현대의 셰프가 폭군으로 악명 높은 왕을 만난다. 그녀는 탁월한 요리 솜씨로 왕의 마음과 입맛을 모두 사로잡을 수 있을까?",
+  "buttons": {
+    "play": { "text": "재생", "icon": "./assets/img_hero/play.png" },
+    "info": { "text": "상세 정보", "icon": "./assets/img_hero/info.png" }
+  },
+  "meta": {
+    "soundIcon": { "src": "./assets/img_hero/sound.png", "alt": "sound_sign" },
+    "ageRating": "15+"
+  }
+}
